\name{criteria}

\Rdversion{1.1}

\alias{criteria}

\docType{data}

\title{ 
  Criteria for the additional conditions
}

\description{
  Criteria to be applied in combination with the conditions. For example, the 
  criteria \code{rangemax} will apply for the first condition (row) in the
  conditions matrix.
}

\usage{data(criteria)}

\format{
  The format is:
  chr [1:3] "rangemax" "rangemax" "min"
}

\source{
  Hortal, J. & Lobo, J.M. (2005) An ED-based protocol for the optimal sampling
  of diversity. Biodiversity and Conservation 14: 2913-2947.
}

\examples{
  data(criteria)
}

\keyword{datasets}

