\name{env}

\Rdversion{1.1}

\alias{env}

\docType{data}

\title{
  Matrix of environmental distances
}

\description{
  A matrix of environmental distances between the 108 sites (with 108 rows and 
  108 columns), summarising climatic characteristics of a 10x10km UTM in the 
  Comunidad Autonoma de Madrid, center of Spain. It was computed using Gower 
  distance.
}

\usage{data(env)}

\format{
  A data frame with 108 observations on the following 108 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{X9}}{a numeric vector}
    \item{\code{X10}}{a numeric vector}
    \item{\code{X11}}{a numeric vector}
    \item{\code{X12}}{a numeric vector}
    \item{\code{X13}}{a numeric vector}
    \item{\code{X14}}{a numeric vector}
    \item{\code{X15}}{a numeric vector}
    \item{\code{X16}}{a numeric vector}
    \item{\code{X17}}{a numeric vector}
    \item{\code{X18}}{a numeric vector}
    \item{\code{X19}}{a numeric vector}
    \item{\code{X20}}{a numeric vector}
    \item{\code{X21}}{a numeric vector}
    \item{\code{X22}}{a numeric vector}
    \item{\code{X23}}{a numeric vector}
    \item{\code{X24}}{a numeric vector}
    \item{\code{X25}}{a numeric vector}
    \item{\code{X26}}{a numeric vector}
    \item{\code{X27}}{a numeric vector}
    \item{\code{X28}}{a numeric vector}
    \item{\code{X29}}{a numeric vector}
    \item{\code{X30}}{a numeric vector}
    \item{\code{X31}}{a numeric vector}
    \item{\code{X32}}{a numeric vector}
    \item{\code{X33}}{a numeric vector}
    \item{\code{X34}}{a numeric vector}
    \item{\code{X35}}{a numeric vector}
    \item{\code{X36}}{a numeric vector}
    \item{\code{X37}}{a numeric vector}
    \item{\code{X38}}{a numeric vector}
    \item{\code{X39}}{a numeric vector}
    \item{\code{X40}}{a numeric vector}
    \item{\code{X41}}{a numeric vector}
    \item{\code{X42}}{a numeric vector}
    \item{\code{X43}}{a numeric vector}
    \item{\code{X44}}{a numeric vector}
    \item{\code{X45}}{a numeric vector}
    \item{\code{X46}}{a numeric vector}
    \item{\code{X47}}{a numeric vector}
    \item{\code{X48}}{a numeric vector}
    \item{\code{X49}}{a numeric vector}
    \item{\code{X50}}{a numeric vector}
    \item{\code{X51}}{a numeric vector}
    \item{\code{X52}}{a numeric vector}
    \item{\code{X53}}{a numeric vector}
    \item{\code{X54}}{a numeric vector}
    \item{\code{X55}}{a numeric vector}
    \item{\code{X56}}{a numeric vector}
    \item{\code{X57}}{a numeric vector}
    \item{\code{X58}}{a numeric vector}
    \item{\code{X59}}{a numeric vector}
    \item{\code{X60}}{a numeric vector}
    \item{\code{X61}}{a numeric vector}
    \item{\code{X62}}{a numeric vector}
    \item{\code{X63}}{a numeric vector}
    \item{\code{X64}}{a numeric vector}
    \item{\code{X65}}{a numeric vector}
    \item{\code{X66}}{a numeric vector}
    \item{\code{X67}}{a numeric vector}
    \item{\code{X68}}{a numeric vector}
    \item{\code{X69}}{a numeric vector}
    \item{\code{X70}}{a numeric vector}
    \item{\code{X71}}{a numeric vector}
    \item{\code{X72}}{a numeric vector}
    \item{\code{X73}}{a numeric vector}
    \item{\code{X74}}{a numeric vector}
    \item{\code{X75}}{a numeric vector}
    \item{\code{X76}}{a numeric vector}
    \item{\code{X77}}{a numeric vector}
    \item{\code{X78}}{a numeric vector}
    \item{\code{X79}}{a numeric vector}
    \item{\code{X80}}{a numeric vector}
    \item{\code{X81}}{a numeric vector}
    \item{\code{X82}}{a numeric vector}
    \item{\code{X83}}{a numeric vector}
    \item{\code{X84}}{a numeric vector}
    \item{\code{X85}}{a numeric vector}
    \item{\code{X86}}{a numeric vector}
    \item{\code{X87}}{a numeric vector}
    \item{\code{X88}}{a numeric vector}
    \item{\code{X89}}{a numeric vector}
    \item{\code{X90}}{a numeric vector}
    \item{\code{X91}}{a numeric vector}
    \item{\code{X92}}{a numeric vector}
    \item{\code{X93}}{a numeric vector}
    \item{\code{X94}}{a numeric vector}
    \item{\code{X95}}{a numeric vector}
    \item{\code{X96}}{a numeric vector}
    \item{\code{X97}}{a numeric vector}
    \item{\code{X98}}{a numeric vector}
    \item{\code{X99}}{a numeric vector}
    \item{\code{X100}}{a numeric vector}
    \item{\code{X101}}{a numeric vector}
    \item{\code{X102}}{a numeric vector}
    \item{\code{X103}}{a numeric vector}
    \item{\code{X104}}{a numeric vector}
    \item{\code{X105}}{a numeric vector}
    \item{\code{X106}}{a numeric vector}
    \item{\code{X107}}{a numeric vector}
    \item{\code{X108}}{a numeric vector}
  }
}

\source{
  Hortal, J. & Lobo, J.M. (2005) An ED-based protocol for the optimal sampling
  of diversity. Biodiversity and Conservation 14: 2913-2947.
}

\examples{
  data(env)
}

\keyword{datasets}

