\name{vectores}

\Rdversion{1.1}

\alias{vectores}

\docType{data}

\title{
  A matrix inlcuding initial localities and additional conditions
}

\description{
  A matrix with all the vectors needed for the example, see definitions of vini
  and conditions.
}

\usage{data(vectores)}

\format{
  A data frame with 108 observations on the following 4 variables.

  \describe{
    \item{\code{BM}}{a numeric vector}
    \item{\code{slope}}{a numeric vector}
    \item{\code{rare}}{a numeric vector}
    \item{\code{usab}}{a numeric vector}
  }
}

\source{
  Hortal, J. & Lobo, J.M. (2005) An ED-based protocol for the optimal sampling
  of diversity. Biodiversity and Conservation 14: 2913-2947.
}

\examples{
  data(vectores)
}

\keyword{datasets}

