\name{vini}

\Rdversion{1.1}

\alias{vini}

\docType{data}

\title{
  A vector with the starting points
}

\description{
  Starting points, those localities known to be well surveyed from previous 
  work.
}

\usage{data(vini)}

\format{
  The format is:
  int [1:108] 0 0 0 0 0 0 0 0 0 0 ...
}

\source{
   Hortal, J. & Lobo, J.M. (2005) An ED-based protocol for the optimal sampling
   of diversity. Biodiversity and Conservation 14: 2913-2947.
}

\examples{
  data(vini)
}

\keyword{datasets}

